pyproject.toml
[project]
name = "images_pipeline"
version = "0.1.0"
description = "A modern Python project for image processing pipeline."
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "boto3>=1.34.0",
    "pillow>=10.0.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "aioboto3>=12.0.0",
]

[project.scripts]
images-pipeline = "images_pipeline.main:main"
process-images = "images_pipeline.process_images:main"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pyright>=1.1.389",
    "ruff>=0.9.3",
    "pre-commit>=4.0.1",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
reportMissingImports = true
reportMissingTypeStubs = false
include = ["src"]
ignore = ["src/images_pipeline/core/image_utils.py", "src/images_pipeline/processors/*.py", "src/images_pipeline/process_images.py", "src/images_pipeline/core/observability.py", "src/images_pipeline/core/services.py", "src/images_pipeline/testing/fakes.py"]

[tool.uv]
package = true

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
    "*/.venv/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod"
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"
